requirements.txt
requests
beautifulsoup4
python-dotenv
pydantic
scikit-learn
joblib
fastapi
uvicorn[standard]
openai
